requirements.txt
asgiref==3.8.1
astroid==3.3.5
<<<<<<< HEAD
boto3==1.35.44
botocore==1.35.44
=======
boto3==1.35.39
botocore==1.35.39
>>>>>>> c6ac06384790f39cd8859a53af52ad612bfa2b30
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cx-Oracle==8.3.0
dill==0.3.9
dj-database-url==2.2.0
Django==5.1.1
django-crispy-forms==2.0
django-environ==0.11.2
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
git-filter-repo==2.45.0
gunicorn==23.0.0
idna==3.4
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.9
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.7.0
